requirements.txt
numpy
pandas
opencv-python
pytorch
torchvision
easydict
matplotlib
pyyaml
tensorboardx

